feat: add queryFields
