--- src/pages/GraphiQlPage/components/Documentation/Characters/index.ts
@@ -1,1 +1,0 @@
-export { Characters } from './Characters';
