+++ src/pages/GraphiQlPage/components/Documentation/ElementDetails/index.ts
@@ -1,0 +1,1 @@
+export { ElementDetails } from './ElementDetails';
