+++ src/pages/GraphiQlPage/components/Documentation/ElementType/index.ts
@@ -1,0 +1,1 @@
+export { ElementType } from './ElementType';
