--- src/pages/GraphiQlPage/components/Documentation/character/index.ts
@@ -1,1 +1,0 @@
-export { Character } from './Character';
